# Refactor checks for sigpanic.