This commit changes the method by which the library checks for the presence
of the runtime `sigpanic` function. Instead of forcing a sigpanic to then grab
a reference to the function, we simply check for the function name. This should
be just as reliable as the alternative and does not cause unwanted interactions
with debuggers.