@dkushner 2017-06-09
Per delve/issues/852.

For reference, this issue is experienced when using delve in with the non-native backend. Essentially, it stems from the fact that LLVM is not able to properly intercept EXC_BAD_ACCESS and so cannot allow the running program to deal with the signal.  Since go-stack uses an invalid dereference to trigger the sigpanic and grab a reference to the runtime function, delve will get stuck in a constant loop rendering it utterly useless.