@ChrisHines 2017-06-11
The link to the delve issue above is broken. Please update your comment to fix the link. Also, can you post [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat) processed before/after benchmarks for this PR? Thanks.