@ChrisHines 2017-07-10
The changes in this PR were rebased on `develop` in 3bd81bdde4686ea4aee961921944d1d012253997 and released in https://github.com/go-stack/stack/releases/tag/v1.5.4